+++ twitter/server/lib/src/entities/explore_feed_entity/entity.dart
@@ -3,22 +3,14 @@ import 'package:horda_server/horda_server.dart';
 import 'messages.dart';
 import 'state.dart';
 
-/// The global entity ID used for the explore feed which all users can see.
-///
-/// This constant defines the single instance ID for the [ExploreFeedEntity]
-/// that allows users discover new tweets, even when the user doesn't follow anybody yet.
-const kExploreFeedEntityId = 'globalExploreFeedEntityId';
-
 /// Represents an explore feed with trending tweets.
+///
+/// This is a singleton entity, automatically created at deployment time.
 ///
 /// {@category Entity}
 class ExploreFeedEntity extends Entity<ExploreFeedEntityState> {
-  Future<ExploreFeedCreated> createExploreFeed(
-    CreateExploreFeed cmd,
-    EntityContext context,
-  ) async {
-    return ExploreFeedCreated();
-  }
+  @override
+  ExploreFeedEntityState? get singleton => ExploreFeedEntityState();
 
   /// For command description, see [AddTweetToExploreFeed].
   Future<RemoteEvent> addTweetToExploreFeed(
@@ -32,11 +24,6 @@ class ExploreFeedEntity extends Entity<ExploreFeedEntityState> {
   @override
   void initHandlers(EntityHandlers<ExploreFeedEntityState> handlers) {
     handlers.addStateFromJson(ExploreFeedEntityState.fromJson);
-    handlers.addInit<CreateExploreFeed, ExploreFeedCreated>(
-      createExploreFeed,
-      CreateExploreFeed.fromJson,
-      ExploreFeedEntityState.fromExploreFeedCreated,
-    );
 
     handlers.add<AddTweetToExploreFeed>(
       addTweetToExploreFeed,